requirements.txt
# HubSpot CRM MCP Server - Dependencies
# Day 11 - December 28, 2025

# Core MCP
mcp==1.25.0

# Async HTTP Client
httpx>=0.27.1

# Environment Management
python-dotenv==1.0.0

# OAuth & Authentication
requests==2.31.0
requests-oauthlib==1.3.1

# Data Validation
pydantic>=2.11.0,<3.0.0

# Date/Time Handling
python-dateutil==2.8.2

# Async Support
aiosqlite==0.21.0  # For potential local caching

# Logging and typing helpers
typing-extensions
